Improve default PE load limit

+++ code/framework/src/launcher/loaders/exe_ldr.cpp
@@ -22,7 +22,7 @@ namespace Framework::Launcher::Loaders {
         hook::set_base();
 
         _origBinary = origBinary;
-        _loadLimit  = UINT_MAX;
+        _loadLimit  = SIZE_MAX;
 
         SetLibraryLoader([](const char *name) {
             return LoadLibraryA(name);
